server: Split people/admin-info from people/index, refactoring
