+++ server/lib/people/index.js
@@ -5,136 +5,12 @@ const { sendMail, updateMailRecipient } = require('@kansa/common/mail')
 const { setKeyChecked } = require('../key')
 const Person = require('./person')
 
-const selectAllPeopleData = `
-  SELECT p.*, preferred_name(p), d.status AS daypass, daypass_days(d)
-    FROM people p LEFT JOIN daypasses d ON (p.id = d.person_id)`
-
 module.exports = {
-  selectAllPeopleData,
-  getMemberEmails,
-  getMemberPaperPubs,
-  getPeople,
   getPerson,
-  getAllPrevNames,
   getPrevNames,
   addPerson,
   authAddPerson,
   updatePerson
-}
-
-function getPeopleQuery(req, res, next) {
-  const cond = Object.keys(req.query).map(fn => {
-    switch (fn) {
-      case 'since':
-        return 'last_modified > $(since)'
-      case 'name':
-        return '(legal_name ILIKE $(name) OR public_first_name ILIKE $(name) OR public_last_name ILIKE $(name))'
-      case 'member_number':
-      case 'membership':
-        return `${fn} = $(${fn})`
-      default:
-        return Person.fields.indexOf(fn) !== -1
-          ? `${fn} ILIKE $(${fn})`
-          : 'true'
-    }
-  })
-  req.app.locals.db
-    .any(`${selectAllPeopleData} WHERE ${cond.join(' AND ')}`, req.query)
-    .then(data => res.json(data))
-    .catch(next)
-}
-
-function getMemberEmails(req, res, next) {
-  req.app.locals.db
-    .any(
-      `
-    SELECT
-      lower(email) AS email, legal_name AS ln,
-      public_first_name AS pfn, public_last_name AS pln
-    FROM People p
-      LEFT JOIN membership_types m USING (membership)
-    WHERE email != '' AND m.member = true
-    ORDER BY public_last_name, public_first_name, legal_name`
-    )
-    .then(raw => {
-      const namesByEmail = raw.reduce((map, { email, ln, pfn, pln }) => {
-        const name =
-          [pfn, pln]
-            .filter(n => n)
-            .join(' ')
-            .replace(/  +/g, ' ')
-            .trim() || ln.trim()
-        if (map[email]) map[email].push(name)
-        else map[email] = [name]
-        return map
-      }, {})
-      const getCombinedName = names => {
-        switch (names.length) {
-          case 0:
-            return ''
-          case 1:
-            return names[0]
-          case 2:
-            return `${names[0]} and ${names[1]}`
-          default:
-            names[names.length - 1] = `and ${names[names.length - 1]}`
-            return names.join(', ')
-        }
-      }
-      const data = Object.keys(namesByEmail).map(email => {
-        const name = getCombinedName(namesByEmail[email])
-        return { email, name }
-      })
-      res.csv(data, true)
-    })
-    .catch(next)
-}
-
-function getMemberPaperPubs(req, res, next) {
-  req.app.locals.db
-    .any(
-      `
-    SELECT
-      paper_pubs->>'name' AS name,
-      paper_pubs->>'address' AS address,
-      paper_pubs->>'country' AS country
-    FROM People p
-      LEFT JOIN membership_types m USING (membership)
-    WHERE paper_pubs IS NOT NULL AND m.member = true`
-    )
-    .then(data => res.csv(data, true))
-    .catch(next)
-}
-
-function getPeople(req, res, next) {
-  if (Object.keys(req.query).length > 0) getPeopleQuery(req, res, next)
-  else
-    req.app.locals.db
-      .any(selectAllPeopleData)
-      .then(data => {
-        const maxId = data.reduce((m, p) => Math.max(m, p.id), -1)
-        if (isNaN(maxId)) {
-          res.status(500).json({
-            status: 'error',
-            message: 'Contains non-numeric id?',
-            data
-          })
-        } else {
-          const arr = new Array(maxId + 1)
-          data.forEach(p => {
-            arr[p.id] = Person.fields.reduce(
-              (o, fn) => {
-                const v = p[fn]
-                if (v !== null && v !== false) o[fn] = v
-                return o
-              },
-              { id: p.id }
-            )
-          })
-          res.json(arr)
-        }
-      })
-      .catch(next)
 }
 
 function getPerson(req, res, next) {
@@ -154,34 +30,6 @@ function getPerson(req, res, next) {
       id
     )
     .then(data => res.json(data))
-    .catch(next)
-}
-
-function getAllPrevNames(req, res, next) {
-  const csv = req.params.fmt === 'csv'
-  req.app.locals.db
-    .any(
-      `
-    SELECT DISTINCT ON (h.id,h.legal_name)
-           h.id,
-           p.member_number,
-           h.legal_name AS prev_name,
-           to_char(h.timestamp, 'YYYY-MM-DD') AS date_from,
-           to_char(l.timestamp, 'YYYY-MM-DD') AS date_to,
-           p.legal_name AS curr_name,
-           p.email AS curr_email
-      FROM past_names h
-           LEFT JOIN log l ON (h.id=l.subject)
-           LEFT JOIN people p ON (l.subject=p.id)
-     WHERE l.timestamp > h.timestamp AND
-           l.parameters->>'legal_name' IS NOT NULL AND
-           name_match(l.parameters->>'legal_name', h.legal_name) = false
-  ORDER BY h.id,h.legal_name,l.timestamp`
-    )
-    .then(data => {
-      if (csv) res.csv(data, true)
-      else res.json(data)
-    })
     .catch(next)
 }
