+++ server/lib/people/router.js
@@ -7,20 +7,56 @@ const log = require('../log')
 const publicData = require('../public')
 const user = require('../user')
 
+const {
+  getPeopleSpaced,
+  getPeopleQuery,
+  getAllPrevNames,
+  getMemberEmails,
+  getMemberPaperPubs
+} = require('./admin-info')
 const people = require('./index')
 const PeopleStream = require('./stream')
 const upgrade = require('./upgrade')
 
 module.exports = db => {
   const router = express.Router()
-  router.use(isSignedIn)
-  router.get('/', hasRole(['member_admin', 'member_list']), people.getPeople)
-  router.post('/', hasRole('member_admin'), people.authAddPerson)
-  router.post('/lookup', isSignedIn, publicData.lookupPerson)
+
+  router.get(
+    '/',
+    hasRole(['member_admin', 'member_list']),
+    (req, res, next) => {
+      const p =
+        Object.keys(req.query).length > 0
+          ? getPeopleQuery(db, req.query)
+          : getPeopleSpaced(db)
+      p.then(data => res.json(data)).catch(next)
+    }
+  )
+
   router.get(
     '/prev-names.:fmt',
     hasRole(['member_admin', 'member_list']),
-    people.getAllPrevNames
+    (req, res, next) =>
+      getAllPrevNames(db)
+        .then(data => {
+          const csv = req.params.fmt === 'csv'
+          if (csv) res.csv(data, true)
+          else res.json(data)
+        })
+        .catch(next)
+  )
+
+  router.membersRouter = express.Router()
+  router.membersRouter.use(hasRole('member_admin'))
+  router.membersRouter.get('/emails', (req, res, next) =>
+    getMemberEmails(db)
+      .then(data => res.csv(data, true))
+      .catch(next)
+  )
+  router.membersRouter.get('/paperpubs', (req, res, next) =>
+    getMemberPaperPubs(db)
+      .then(data => res.csv(data, true))
+      .catch(next)
   )
 
   const peopleStream = new PeopleStream(db)
@@ -30,6 +66,9 @@ module.exports = db => {
       else peopleStream.addClient(ws)
     })
   })
+
+  router.post('/', hasRole('member_admin'), people.authAddPerson)
+  router.post('/lookup', isSignedIn, publicData.lookupPerson)
 
   router.use('/:id*', user.verifyPeopleAccess)
   router.get('/:id', people.getPerson)
@@ -48,11 +87,6 @@ module.exports = db => {
     hasRole('member_admin'),
     upgrade.authUpgradePerson
   )
-
-  router.membersRouter = express.Router()
-  router.membersRouter.use(hasRole('member_admin'))
-  router.membersRouter.get('/emails', people.getMemberEmails)
-  router.membersRouter.get('/paperpubs', people.getMemberPaperPubs)
 
   return router
 }
